refactor(product-details): update import paths and introduce new context for targets
- Refactored import paths in product-details and related components to align with the new structure.
- Introduced a new context provider for managing product detail targets, enhancing the extensibility of the product details component.
- Updated the ProductDetails component to utilize the new targets system, replacing deprecated props.

These changes improve the organization of the codebase and facilitate easier customization of product detail rendering.
